Create custom tabs navigator
* Refactoring and move of some components.
* Tabs navigator created and exemple use.
