+++ package.json
@@ -29,6 +29,9 @@
   "dependencies": {
     "@hookform/resolvers": "^2.9.3",
     "@react-native-async-storage/async-storage": "^1.17.7",
+    "@react-navigation/core": "^6.2.2",
+    "@react-navigation/native": "^6.0.11",
+    "@react-navigation/stack": "^6.2.2",
     "@reduxjs/toolkit": "^1.8.3",
     "expo": "~45.0.0",
     "expo-constants": "^13.1.1",
@@ -46,6 +49,8 @@
     "react-native": "0.68.2",
     "react-native-gesture-handler": "~2.2.1",
     "react-native-reanimated": "^2.9.1",
+    "react-native-safe-area-context": "4.2.4",
+    "react-native-screens": "~3.11.1",
     "react-native-web": "0.17.7",
     "react-redux": "^8.0.2",
     "redux-logger": "^3.0.6",
