+++ src/app/Navigation.tsx
@@ -1,0 +1,3 @@
+import { createStackNavigator } from "@react-navigation/stack";
+
+const Primary = createStackNavigator();
